Pulling Host from URI when necessary.
This commit pulls a Host header from a URI if the messsage does not have
an explicitly set Host header. When a Host header is inferred from a
URI, the Host header is marked as "soft". Subsequent "withUri" method
calls can overwrite a Host header if the header is marked as "soft".

This commit also makes the Host header first in the list of headers as
per http://tools.ietf.org/html/rfc7230#section-5.4.

For context on pulling the Host header from the URI, see
https://github.com/php-fig/fig-standards/commit/9040c10e356322ff9f3d2ebaffa03b4f6847785f